refactor: properly namespace shared code
