+++ src/main.cpp
@@ -1,9 +1,4 @@
-#include "http/RequestParser.hpp"
-#include "http/ResponseParser.hpp"
-#include "http/http.hpp"
-#include "shared/stringUtils.hpp"
+#include "shared/log/Logger.hpp"
 
 int main() {
-	size_t foo = shared::string::toNum<size_t>("1234", std::hex);
-	std::cout << foo << std::endl;
 }
